Flatten constants
